Move MAC spoofing failure messages in systemd units (refs: #17813)
Add 2 systemd --user unit templates (config/chroot_local-includes/usr/lib/systemd/user/tails-mac-spoof-failed-panic-{success,failure}@.service), which:
  - are not enabled by default in the 52-update-rc.d hook
  - take over the job of displaying the failure notification (what's done respectively in notify_panic_success() and notify_panic_failure())
  - are WantedBy=desktop.target
  - are parameterized by the low-level name (e.g. eth0) of the network card whose MAC address could not be spoofed, i.e. calling get_name_of_nic, so that if this part fails, it cannot possibly affect the MAC spoofing panic code path itself: it can only affect displaying the notification
On MAC spoofing failure, tails-spoof-mac instantiates and enables the one relevant unit template among those 2, with the name of the relevant network card as the parameter ⇒ when the amnesia user's systemd --user starts desktop.target, it will start this unit, which will display the notification.

Contrary to what was proposed on #17813, computation of the name of the network card was kept in the main script as it relies on the module presence.
Thus, the full name of the interface is passed through a /run/tails-spoof-mac-notify-${nic} file.