Do not use proprietary keybox format to exchange trust roots.
Previously, the archive keyring was imported into an ephemeral GnuPG
state.  Then, some private GnuPG state is extracted and used as input
to apt-secure.

This step was introduced without motivation in the commit message.
The one thing it does do (or what we nowadays rely upon) is to convert
the input material which contain two blocks of ASCII-armored OpenPGP
certificates (a somewhat supported but non-standard thing) into a
single ASCII-armored OpenPGP keyring (the standardized interchange
format).

Note how previously the key material was taken from GnuPG's private
state.  With modern GnuPG, the keyring is actually stored in a
different, proprietary format called a key box.  I assume that using
this format with apt-secure works only accidentally because apt-secure
also uses GnuPG.

So, add an a posteriori justification for what is happening, and
use gpg --export to export it to a standard keyring.