Improve test.
Previously, if no keys were found, the test step would succeed.  But
this is a vacuous truth.  Instead, assert that there are OpenPGP and
APT keys.  I feel like this better captures the intention of the test:
to make sure that you can authenticate peers in N months from now.