Implemented a Python wrapper for the Zbar GTK widget.
Zbar offers a GTK widget, and it used to have a Python wrapper.
However, it was for Python 2 only.
With Python 3, GTK 3, etc, there is a new infrastructure for
GObject-based structures (e.g., all GTK widgets, including the Zbar
one).
So, once the widget is somehow created and returned to Python, gi does
some magic for which it is automatically recognized and all its methods
and properties become accessible.
The only exception is zbar_gtk_request_video_size, for which I have
added a wrapper. I could not find how to register it as a custom method
for the ZBarGtk objects, so it accepts one as a first argument.

See the attached example.py script for further information on the
usage.