Flatpak runtime: Mount / into flatpak runtime instead of the SquashFS
The initial implementation of our Flatpak runtime mounted the SquashFS
(filesystem.squashfs + incremental upgrades) and used that as the root
filesystem (similar to how we set up the runtime of the Unsafe Browser).

On the one hand, that provides a deterministic and clean environment,
which also has a security benefit: That environment is guaranteed to not
contain any secrets which a compromised app could leak (regardless of
AppArmor policies).

On the other hand, it's surprising during development that changes made
to the host's root filesystem (for example by systemd services) are not
visible for apps running in our Flatpak runtime.

We haven't clearly decided yet which approach we prefer, but we want to
start with the more "traditional" and less surprising approach to make
the host's root filesystem accessible to the apps running in our Flatpak
runtime.