Cloner: unmount all partitions related to the target when needed
No partition (e.g. "siblings") on the target device can be mounted (or
unlocked, if a LUKS device) while partitioning or installing the boot
loader, so we have to deal with that.

This commit rewrites unmount_device() completely to make it act on
up-to-date information (instead of possibly outdated and often
completely incorrect information stored in `data['mount']` and
`data['mounted_partitions']`, the latter which we ripped out
completely) and also deal with locking of LUKS-encrypted volumes. The
new boolean flag `unmount_all` was added to also deal with _other_
partitions on the target device.

Refs: #20149