Fix dump-user-env
We handle cases that some of the environment variables of ENV_VARS_TO_DUMP
are not set. Currently, those conditions will never be met, because if any
of the env vars in ENV_VARS_TO_DUMP are not set, `os.environ[e]` will
fail with a `KeyError`.

Using `os.getenv(e)` instead fixes that.