Test suite: lower needed pattern coverage after filling the memory
We do this in the 'Tails erases memory on DVD boot medium removal:
vfat' because we are seeing this fail there for reasons we cannot
really understand, especially since it doesn't happen in the very
similar 'Erasure of read and write disk caches on unmount: vfat'
scenario.

I cannot come up with a reason to justify investigating this deeper,
all we want to do with the pattern coverage check is ensuring that
there are "many" patterns before the wipe and check that there are
no/few patterns remaining after the wipe. 90% is sufficient for this
"anti-test".

Refs: tails/tails#20156