tails-iuk-generate-upgrade-description-files: only warn about missing --previous_version
While releasing the first Tails X.0 pre-release that gets security
support (typically X.0~betaN or X.0~rc1), it's expected that the RM passes
no --previous_version argument, so let's make it optional.

However, this is a rare situation, so when no such argument was passed, there's
a chance this was a mistake, so let's warn about it. Granted, that's not very
loud, but we have a safeguard in case this was a mistake: manual QA will fail.

refs #20197