Make QT apps shipped in Tails use cursor size configured in GNOME
Commit b849c4973db669216eacfcc19675a337c0c0ab30 sets XCURSOR_SIZE
globally in /etc/profile.d. That has the downside that the cursor
size is not updated when the user changes the cursor size in the
GNOME settings.

This commit sets XCURSOR_SIZE to the value configured in GNOME when
the app is started. If the user changes the cursor size after the app is
started, the app needs to be restarted for that change to take effect.

b849c4973db669216eacfcc19675a337c0c0ab30 is still useful because it
makes other QT apps which the user installs use GNOME's default.

Closes #20206