Disable Tracker
The tracker-extract-3.service often fails to connect to the
tracker-miner-fs-3.service, causing a timeout which has the result that
the desktop is shown 30s later than usual.

To the best of my knowledge, it's not used for anything in Tails. The
one thing I know of where it's usually used is the file search in the
GNOME Shell overview, but that's disabled by default in Tails.

Closes #20220