Cloner: improve detection of persistent storage and allow installing to LUKS devices
First of all, this logic belongs to creator.py, not the GUI code as it
was previously.

The previous logic was a bit simplistic, assuming that any LUKS device
is a persistent partition (if I understood it correctly). For that
reason we also prevented users from installing over a LUKS device,
which is an surprising limitation.

However, it should be mentioned that the real motivation for this
commit is that the recent work that make the Cloner "only list
partitions as targets when they are upgradeable" broke the old check
for detecting if persistence is present on the target (used for
selecting the wording for the confirmation dialog) and wouldn't be
easy to fix.