tps: Fix method name included in log output
By default, the thread name is "Thread-$target", where $target is the
name of the target function. That thread name, without the "Thread-"
part, is used as a prefix to log messages printed by the thread (see
config/chroot_local-includes/usr/lib/python3/dist-packages/tps/logging.py).

The intention was to have the D-Bus method name in the log messages, but
we don't set the target function to the D-Bus method but to
`handle_method_call`.

This commit explicitly sets the Thread.name to the D-Bus method name.