tps: Reconnect to udisks if the D-Bus connection is lost
I've seen the case that udisks was oom-killed and then a subsequent call
to ChangePassphrase failed because the `udisks.get_object` call returned
None.

We now automatically create a new D-Bus connection to udisks (which
causes udisksd to be spawned via D-Bus activation) in case that the
connection is lost.