Firewall: don't DROP OUTPUT, LOG and REJECT instead
When we flipped from allowlist to blocklist to for amnesia we
implemented "block" as just DROP, not our usual LOG and REJECT as the
catch-all in the OUTPUT chain, which I think was an oversight.

This will be important for an upcoming automated test where we want to
verify that packets are dropped when they are supposed to (and vice
versa).