Silence linting failure
If we checked for errors, on failure, there's little we could do to make the UX
better: the user won't see the desired documentation appear anyway.

The best we could do is to encourage them to report the bug to us.
If 8a95b0894c960ce3c22946bd70e359a644d3df3a is the simplest way to do that, then
I'm not convinced it's worth it, and ENOTIME to review & test this in time for
6.1.