Cloner: appease Ruff vs PLW2901 (redefined-loop-name)
In this instance redefining the variable does make sense, since we are
looping over object that themselves might not be the final object we
want to interact with, as in the case of LUKS volumes.

If we wanted to fix this any way with:

    for obj in unmount_candidates:
        volume = obj
        # Use `volume` instead of `obj` everywhere below
        [...]

we don't actually fix anything. So let's just silence Ruff in this
case.