Update random seed after the GNOME session has started
This finally fixes the race condition of the
"A random seed is read during boot and written during boot and shutdown"
scenario.

The reason it was flaky is that we read the random seed once the Welcome
Screen has started and sometimes the tails-update-random-seed-sector.service
has already started at that point. It had a After=multi-user.target
dependency but that target is already reached when the Welcome Screen
has started.

There was no systemd system target which is reached after the Welcome
Screen, so this commit adds one which is only reached once the GNOME
session has started.