Start root terminal as part of app-org.gnome.Terminal.slice
Before, the terminal was started in the cgroup of
org.gnome.Shell@wayland.service, which means that when it's killed by
systemd-oomd, Wayland is killed as well, resulting in the whole GNOME
session to be stopped.

By starting gnome-terminal as amnesia and only changing the user to root
inside the terminal causes it to be put in it's own cgroup, so that only
the gnome-terminal instance is killed by systemd-oomd.