auto/config: don't set TAILS_GIT_BASE_COMMIT when building from tags
Bugfix on ca1efbb01063aca434788e2bf09077b60a3a8fd7 introduced in Tails
6.0~rc1.

The condition was `[ "$GIT_BRANCH" != "$(base_branch)" ]`, and when on
a tag `$GIT_BRANCH` is unset, but `base_branch` will still output the
contents of `config/base_branch` (e.g. `stable` for 6.1), so it
evaluates to `true` and `TAILS_GIT_BASE_COMMIT` is written, which
breaks reproducibility if your origin/stable isn't in the same state
as for the RM/jenkins when they built the release, which is very
likely.

While we're at it we also drop TAILS_GIT_BASE_BRANCH since it's only a
concept that is relevant when working with branches, so it doesn't
make sense for releases.

Refs: tails/tails#20311