tps-frontend: Hide unexpected error messages behind a "Details" expander
Some unexpected error messages returned by the backend might be
confusing to the user. An example is the error message in the
description of #20163:

  gi.repository.GLib.Error: udisks-error-quark: GDBus.Error:org.freedesktop.UDisks2.Error.Failed: Error creating partition on /dev/sdb: Process reported exit code 2:
  Warning! Main and backup partition tables differ! Use the 'c' and 'e' options
  on the recovery & transformation menu to examine the two tables.

  Warning! One or more CRCs don't match. You should repair the disk!
  Main header: OK
  Backup header: OK
  Main partition table: OK
  Backup partition table: ERROR

  Invalid partition data!
   (0)

There is nothing actionable for the user in that error message and it's a
lot of text which is incomprehensible and confusing for most of our
users. Instead of dumping all that in the error dialog, we hide it
behind a "Details" expander now.