Test suite: split code for improved organization
We must `require 'features/support/monkeypatches/extra_hooks.rb'` in
`features/support/hooks.rb` since it references things defined
there. We didn't have to do this before because the alphabetic sorting
made `extra_hook.rb` load before `hooks.rb`, but that is not the case
any more. This would not have been an issue if cucumber used `require`
(which doesn't reload code) when loading the step definitions and
support code, but it uses `load` (which allows reloading).