Simplify tails-unblock-network
Instead of asynchronously starting a systemd service and waiting for
its commands to take effect, just run those commands in the
tails-unblock-network script itself.

The now removed tails-unblock-network.service did not pull in any
dependencies and it was not used anywhere else than in the
tails-unblock-network script, so it should be safe to remove it.