Test suite: make sure we only apply monkeypatch once
Originally I just got warnings that the constants defined in
`extra_hooks.rb` were modified and thought to only guard them, but I
don't think some of the monkeypatching in this file is safe to apply
twice, so let's hide it all behind a guard.

BTW, this feels like C-style header guards. :P