Start root terminal in its own systemd scope
Starting gnome-terminal as amnesia had the drawback that new windows /
tabs were run as amnesia instead of root.

We now run root terminal in its own systemd scope, which still avoids
that gnome-shell is killed when a process in a root terminal uses too
much memory.

A downside is that all tabs / windows started from the root terminal are
killed as well. We could avoid that by configuring gnome-terminal to run
the following custom command instead of the shell:

   systemd-run --quiet --scope -- bash

That needs to be configured via dconf, which I found too much effort for
now.