tails-unblock-network: Terminate udevadm trigger after 120 seconds
We used `udevadm settle` before, which had a default timeout of 120
seconds. `udevadm trigger --settle` does not have a timeout, so it could
potentially block forever if any of the triggered events never finishes.
To avoid that we block the boot forever, we use the `timeout` command to
send SIGTERM to `udevadm` if it takes longer than 120 seconds.