Verify and repair GPT after first boot repartitioning
We've seen reports where both the MBR and the GPT were corrupted after
the repartitioning, which caused the Persistent Storage partition to
fail to be created.

We now verify and try to repair the GPT (and protective MBR) up to three
times before either continuing or telling the user to use a different
USB stick.