Release process: drop the need for re-tagging (refs: tails#20314)
By reordering how we do things we don't need the now-removed "Tag the
release in Git" section where we created a tag that later had to be
overwritten, is is fine to do "Prepare the versioned APT suites" after
the "final" tag (now the only tag). So no need for re-tagging, yay!

I will test this during the 6.2 release, let's see how it goes!