Test suite: Fix duplicate drive address for disk
Tests which plugged more than three disks failed with:

    And I plug SATA drive "live_hd"
      Call to virDomainDefineXML failed: unsupported configuration: Found duplicate drive address for disk with target name 'hdc' controller='0' bus='0' target='0' unit='2' (Libvirt::DefinitionError)

That's because the disks were assigned the device "sdc", which conflicts
with the "hdc" device which we hardcoded for the cdrom drive.

It's fixed by assigning the cdrom drive a device letter which comes
later in the alphabet. Note that we can't assign it the device name
which the guest OS assigns to the device (sr0), because libvirt doesn't
support that name.