Report any error during first boot repartitioning
Previously we only checked if the GUID was randomized. We saw error
reports which showed that the backup GPT was corrupt, which caused
fatresize to fail and the Persistent Storage partition to fail to be
created.

We now show the "Something went wrong when starting your Tails USB stick
for the first time" message if any error occurs during first boot
repartitioning or if the GPT is corrupt or the system partition not
correctly resized.