Test suite: stop Tor later in After hook
This way the resulting journal entries are not included in the Tor
logs artifact, which otherwise could be confusing when reading
them. It's now also not included in the full journal, but what value
could it possibly add?

While we're at it, let's ensure we stop Tor even if there are uncaught
exceptions during the After hook, since not doing so can interfere
with the next scenario.