Cloner: reduce complexity by moving code out of nested if
The reviewer rightfully pointed out that it was messy.

Moving the passphrase confirmation for when cloning persistence out of
the Install case (self.opts.partition == True) also prepares for
making it possible to clone persistence in the Upgrade case, which
makes it faster to backup your persistent storage, especially together
with tails#20362.

Until then, when in the Upgrade case (self.opts.partition == False)
self.opts.clone_persistent_storage_requested should always be False so
the dialog isn't uselessly showed: see
on_check_button_clone_persistent_storage_toggled(), which is the only
place we change the state of that variable, and how it calls
update_start_button(), which makes it impossible to reach a
problematic state.