Cloner: drop useless variable and the nested if-statement it's used in
This variable is something we inherited from liveusb-creator.

The only time it would matter is if it's set to True before calling
begin(). Maybe that can be the case if there is a failure during a
previous call of begin() and the installation is aborted, but we try
again, which sounds like something we absolutely don't want. Die!