Fix issue with partition table verification
The verification failed in many of our test suite scenarios, because
sgdisk created the system partition with quite a few sectors less than
expected (3937 sectors less in the cases I manually inspected).

It's not clear to me why sgdisk does that (maybe something to with
alignment, but the size difference seems too high to be explained only
by that), but lets account for it in the verification code.