fixup: Fix scenario "Downloading files with the Tor Browser"
The file chooser dialog opened by the Tor Browser in this scenario has
the name "Enter name of file to save to…" instead of "Save As". To avoid
unnecessary complexity, this commit removes the name argument and makes
`desktop_portal_save_as_dialog` return any "file chooser" child of the
"xdg-desktop-portal-gtk".