Cloner: Move tps_totalsize variable back to ProgressThread
The purpose of ProgressThread is to monitor the progress of the copy
operations, including copying the files on the tps partition, and
updating the progress bar accordingly. It makes sense that it checks the
total size of the tps partition itself instead of receiving it in an
accumulated "size" argument.