Cloner: Remove useless division by 1024
It doesn't make sense to divide both the dividend and the divisor of the
progress calculation by 1024. It might make sense if we rounded the
result, which would have the effect that the progress updates are more
coarse, i.e. per changed KiB instead of per changed byte. It would be
unexpected for the copy operation to copy less than 1 KiB per 0.1 second
cycle though, so I won't make that change and instead remove the
division by 1024.

Co-authored-by: Ben Westgate <benwestgate@protonmail.com>