Update the master branch with updates from IkiWiki
This commit turns on IkiWiki's "revision control system" option for GitLab CI
jobs that run on the master branch, effectivelly causing IkiWiki to commit and
push IkiWiki updates back to the repository.

This is a replacement for the same process that was previously done in our web
VM using IkiWiki's CGI facilities and its pingee plugin.

refs sysadmin#17364