tps-frontend: Fix opening documentation causing window to freeze
Since af7421d971f20be7d267ecc8fde3fb34c3321721 we capture the stderr of
the tails-documentation command (to show it in the error message if the
command fails). That causes the `subprocess.run()` call to block until
the stderr file descriptor was closed. Because tails-documentation calls
gtk-launch which in turn passes the stderr file descriptor on to the
launched application, the file descriptor is not closed until the
launched application is closed.

This commit uses systemd-cat to connect the stderr (and stdout) of
gtk-launch to the journal instead, which has the effect that the stderr
file descriptor of tails-documentation is not passed on to gtk-launch
and is therefore closed immediately when tails-documentation exits.