Repair first-boot partitioning: write the random seed only up to 1000 times during first boot
Quite often, when we write it 1000 to 2000 times, this breaks first-boot
partitioning: "udevadm settle" fails with "Timed out for waiting the udev queue
being empty". In turn, this breaks creating a Persistent Storage and will break
automatic upgrades sooner or later.

We had no report of such bugs when we write fewer than 1000 times, so let's
lower the upper bound to 1000.

Whether 1000 is enough to obfuscate how many times Tails has been booted
is unclear: it's unclear why 2000 was chosen initially. As per the
corresponding design doc, this is an best effort attempt anyway, as it's
known to not work in some cases.

refs #20451