Arti: start via NetworkManager hook instead of during boot
I tried to set the unit's dependencies so it only starts after the
network is online, but that is not what network-online.target does,
and we start NetworkManager-wait-online.service manually, so arti just
kept on starting too early, causing it to spam the log with:

    WARN tor_chanmgr::transport::default: Connection to [scrubbed] failed: error: Network unreachable (os error 101)

and put Arti into a waiting state before it tries again, delaying
bootstrap.

Any way, this is closer to what we do with tor, so probably a good
idea any way. The above warning is sometimes logged once of twice, but
there is at least no delay bootstrapping.