Test suite: Fix test of Persistent Storage when GPT backup is corrupt
Corrupting the MBR as well causes udisks to detect no partition table on
the device, resulting in the generic "Sorry, it is impossible to create
a Persistent Storage on this device." message to be shown in
tps-frontend.

Corrupting only the GPT backup table causes fatresize to fail but still
allows udisks to detect a partition table, therefore resulting in the
expected "resizing the Tails system partition failed" message being
shown in tps-frontend.