Simplify
As discussed on
https://gitlab.tails.boum.org/tails/tails/-/merge_requests/1590#note_238812:

Advantage: simpler, no need for such ungodly shell programming hacks.

Drawback: no error is reported to the user when we hit 1 of these failure modes;
but hitting these failure modes should be super rare for end-users, in practice
they are only common while working on these scripts; and it's not as if the
existing error messages were particularly helpful or actionable anyway (beyond
"no need to wait any further"); hopefully they'll report a bug and we can figure
out what went wrong from the Journal.