Don't try to localize strings in non-existent files
This file does not exist in Tor Browser 13.5.1 in en-US locale.
We've already noticed that it was not used anymore in some places
(01add28e3dbb3ba5c49a4a37518c80288c1c3953), it seems it is getting deprecated.
Either way, the command I'm removing here fails on Tails 6.5, and despite that
the Unsafe Browser has the correct name in the few important places I could see.