Revert "Update the Vagrant basebox for any change under vagrant/."
This reverts commit 3fc2a5be04ae7f9781c053f00ebbfeeec88a549f.

The reasoning in that commit is at least currently wrong[0]: the only
files inside ./vagrant that affect base box generation are:

* those in the vagrant/definitions/tails-builder directory.
* vagrant/lib/tails_build_settings.rb: The box_name() method is called
  for the base box filename, but changing how that name look will in
  itself ensure that a new base box is built since it was missing (see
  the logic in Rakefile).

In particular, vagrant/Vagrantfile is only used when vagrant creates
the transient VM used for building. That has no impact on base box
generation at all, it *uses* (makes a copy of) an already generated
base box, and that's the only relationship between them.

[0] It has not been true since we started using vmdebootstrap. Maybe
(didn't look carefully) it was true 8+ years ago and earlier when we
used Veewee to generate base boxes.