Build system: consolidate memory sanity checks into a task that vm:up depends on
Vagrant outputs "cannot set up guest memory 'pc.ram': Cannot allocate
memory" if there isn't enough memory to start the VM, but it is buried
in a huge error message and difficult to spot. So it's best if we do a
sanity check and write our own error message.

While we're at it we also remove the VM memory check done in the build
task: in 2017 we adopted a new paradigm where we "provision a new
Vagrant VM for every ISO build" and it might have been useful before
that. I don't think it does anything useful now, and I cannot recall
seeing that check fail. I think the only way you currently can trigger
it is if you do something silly like:

    TAILS_BUILD_OPTIONS=noram rake vm:up
    TAILS_BUILD_OPTIONS=ram rake build