Manual QA: remove not-super-necessary test
refs https://gitlab.tails.boum.org/tails/tails/-/merge_requests/1656#note_241817

Initially, we thought our automated testing of WebRTC being disabled was too
weak to fully remove the manual test. Indeed, those tests totally rely on the
output of some 3rd-party websites, that we may or may not trust.

Let's save some manual QA time: let's trust both Tor Browser upstream and us to
not enable this setting by mistake, and even if they do, we would be really
unlucky if on top of that, the 3rd-party websites we used were broken
or dishonest.