Test suite: configure the Chutney client node to launch after the Tor network is bootstrapped
The default launch_phase is 1, which we keep for authorities and
relays, i.e. the actual Tor *network*, so they are bootstrapped and
the consensus is published for launch_phase 2, when we start the
client (and bridge nodes) so it won't get stuck trying to bootstrap
when the network isn't ready.