tps: Improve password strength meter
Change how we calculate the score which we base our password strength
meter on.

Use the python3-zxcvbn package instead of pwscore to calculate a score.
That package seems to a lot better at detecting passwords which are
likely to be on password lists, like "password1234!", and accordingly
gives them a low score.

This change also tries to better match the feedback provided by the
strength meter with our recommendation to choose a passphrase of 5 to 7
random words, so that such a passphrase should be considered "strong"
while a shorter passphrase is considered no better than "fair".