Only kill GDM when I/O errors are detected before starting the Welcome Screen
This script will be run by Debian-gdm when the Welcome Screen starts,
and by the Live user when logging in. We only want to kill GDM in the
first case, since that indicates that I/O errors were encountered
while touching the stuff necessary to boot. If the I/O errors occur
during the Welcome Screen it likely is when attempting to unlock the
Persistent Storage, and then we want them to be able to log in so we
can show them recovery instructions.