Welcome Screen: drop unnecessary code
language_name in self.language_names_per_language was set through
self._language_name() so the removed line did nothing except violating
PLW2901 (overwrite loop variable).

Looking at the (out-of-tree) Git history of the Greeter it seems to
just be an oversight: the look used to be:

    for lang_code in lang_codes:
        language_name = self._language_name(lang_code)
    [...]

so when we moved to self.language_names_per_language that line was
probably forgotten to be removed.