NetworkManager: enable ip6-privacy for all connections (refs: tails#20478)
This randomizes the IPv6 address (except the network prefix) using the
method described in RFC8981 so we don't leak the MAC address through
EUI-64.

It is worth noting that for whatever reason Tails didn't seem to use
EUI-64 before setting this, but now we're at least sure what method is
used, and that it is safe, and that whatever default we were using
doesn't suddenly change to EUI-64.

ipv6.ip6-privacy=2 means that NM will set the interface's use_tempaddr
to the same value, see the following for details:

  https://www.kernel.org/doc/Documentation/networking/ip-sysctl.txt