Only look at real ethernet NICs when considering MAC spoofing panic notification message
The previous code was actually a bit incorrect: our veth-* interfaces
will all become just "veth". But fixing it will make it so we always
have "other" wired NICs, since ethtool returns 0 for them. So the old
code's mistake incidentally fixed that since "veth" doesn't exist. But
the real fix is to look at interfaces for real devices, which we
already have code for (and then we can also remove the special case we
have for the loopback device).