Welcome Screen, tpsd: drop forceful_fsck parameter
It is not used for anything except keeping track of whether we already
did a manual repair in the Welcome Screen. But using it for that made
it so that after the user requested a forceful repair, all fsck:s done
in the tpsd backend when trying to mount the device again also are
forceful, which violates the design. So on the Welcome Screen-side we
use another variable to keep track of this.