Test suite: avoid GNOME notification race condition (refs: tails#20286)
This is very similar to commit
c8cdbf01d6a8711895bf790c9b991e271f9df27e for tails#14596. I'll paste
the relevant part of intrigeri's commit message below:

> [...] sometimes "all notifications have disappeared" (which actively
> clears the list of notifications) hides the very notification we're
> looking for and want to interact with. Whenever this happens, 'I see
> the "The installation of your additional software failed"
> notification after at most 300 seconds' will succeed anyway (Dogtail
> will find notifications even after we've cleared them, go figure)
> but thankfully 'I can open the Additional Software log file from the
> notification' will fail to interact with an invisible
> notification (anything else would be utterly confusing).
>
> To avoid this problem, we need to stop clearing all
> notifications. But then the "Warning: virtual machine detected!"
> notification may hide the one we're looking for. So let's disable
> the service that would display this spurious notification.

Nice!

However, there is also this:

> Given this scenario runs Tails offline, this should be the only
> notification displayed at login time, so this blocking it should be
> a sufficient substitute for clearing all notifications.

This is a different scenario, and in this part of it the network is
connected, so there are a few more notifications. However, from my
testing, if there is a queue of notifications GNOME will display each
of them for a while, in order, so eventually it will show the one we
want, and ours should be the last one so it should even stay
longer.

The key here, I think, is to not run the "all notifications have
disappeared" step. Let's try it!