needed-package-updates: don't skip tons of relevant packages
For whatever reason, for tons of our packages, within this loop, as expected
pkg_dict.get(p) returns None, because we've not seen the package yet (which is
the most common case), *but* pkg_dict.get(p) > v is true, so we skipped that
packaged and did not add it to pkg_dict.

To fix this, let's only do the comparison when pkg_dict.get(p) is True,
so we only compare non-None versions.