needed-package-updates: prevent future issues with Version comparisons
Apparently None > debian.debian_support.Version("1") is True, which
resulted in a serious bug (which prevented several outdated packages
from being reported) that was fixed with tails!1733.

In order to prevent this issue in other places (or being reintroduced
in the future) we make all our Version comparisons bail unless
compared with another instance of the same class.