Release process: fix building IUKs with sid's default tmpfs /tmp settings and 32GB RAM
Nowadays Debian sid defaults to a tmpfs on /tmp whose size is 50% RAM. So with
32GB of RAM, a Release Manager could not build IUKs with the command we had in
our release process. They had to pass a custom --tmp-dir to
wrap_tails_create_iuks.

To fix this, let's manage a dedicated temporary directory under /var/tmp, that
has more chances to have enough space; and set this up in a way that it gets
cleaned up during boot, to avoid filling the disk.

refs tails/tails#20604