Install intel-microcode from stable, not from sid
Until Tails 6.8.1 we were shipping intel-microcode 3.20240813.2 from sid, and
with this change we'll get the same version backported to
stable (3.20240813.1~deb12u1).

https://tracker.debian.org/pkg/intel-microcode now gets regular backports to
stable, be it via proposed-updates or via security, depending on the urgency.

So let's reduce our churn and drop the pinning that pulls this package from sid,
in favor of getting the updates from the various stable-* suites that we have
configured. In other words, I think it's reasonable to trust the maintainer to
figure out which changes are worth backporting to older Debian (that we're based
on) and when, and we can stop duplicating their work (which is essentially what
we currently have to do to decide "do we update this package or not" on an issue
such as this one). To be clear, this implies we'll get some security fixes
later, e.g. 3.20240910.1 is in sid since a few weeks, and fixes some
vulns (https://bugs.debian.org/1081363) that are not fixed in stable yet. It's
figuring out precisely this stability/robustness vs. security balance that
I would prefer we delegate to the maintainer of the package.