Test suite: wait for persistence to be fully activated instead of merely unlocked (refs: tails#20619)
In #20619 we have seen that Dogtail sometimes fails to find the
Greeter when greeter_language() is called by the `persistence is
successfully enabled` step, due to the Greeter freezing during the
unlock operation. This is only an issue because we earlier in that
step only waited until the persistent storage was unlocked, not fully
activated, and the time between those can be more than the 30 seconds
that greeter_language() waits for the Greeter to become unfrozen
again. So let's instead wait for persistence to be fully activated
before calling greeter_language().

PS: It would be nice to fix the freezing issue in the Greeter, but I
have no clue how to do that at the moment.