Upgrader: send more output to the Journal to debug #20593
If #20593 happens again, this will help us tell with certainty in which code
branch we are.

If this happens again we should consider adding even more debug logging (e.g.
7ba79fbb10235fe3b1cf0427072d45eb86dd22ff) but this would be extremely noisy in
the Journal. Too bad this is probably a race condition, so "try again with debug
mode enabled" is not a viable approach.